feat(V1.5-S0): export Databricks→DuckDB real, repo actualizado, bootstrap R2
Exporta 15/16 tablas gold+silver desde Databricks a JSON,
construye DuckDB con tipos reales via DESCRIBE, valida endpoint
sales-summary con 0-diff vs raw_responses.json.

Bootstrap del API: si /tmp/motoshop_gold.duckdb no existe,
descarga automaticamente desde Cloudflare R2.

Curl evidence (API local port 8001, DATA_BACKEND=duckdb,
archivo descargado de R2):

$ curl -H "Authorization: Bearer <token>" "http://localhost:8001/api/metrics/sales-summary" | jq '.ventas_mes_actual, .num_facturas'
23516508.33
911

Diff vs raw_responses.json#SALES_SUMMARY: 0 (tolerancia 0 exacta)
DuckDB: out/motoshop_gold.duckdb (7.3 MB, 105,449 filas)
R2: motoshop_gold.duckdb subido ✅
Bootstrap: descarga 7.3 MB desde R2 en cold start ✅

Archivos:
- pipeline/export_databricks.py: exporta tablas Databricks a JSON (env vars)
- pipeline/build_duckdb_from_export.py: construye DuckDB con tipos reales
- motoshop-app/api/src/motoshop_api/metrics/repo_duckdb.py:
  usa tablas reales + _bootstrap_duckdb_from_r2()
- motoshop-app/api/render.yaml: env vars DATA_BACKEND, DUCKDB_PATH, R2_*
- motoshop-app/api/pyproject.toml: duckdb>=1.0, boto3>=1.34
- pipeline/spike_sales.py: DECIMAL(18,2) para tolerancia 0

+++ pipeline/spike_sales.py
@@ -133,8 +133,8 @@ def build_duckdb(duckdb_path: str | Path) -> str:
             nom_producto VARCHAR,
             cod_bodega VARCHAR,
             nom_bodega VARCHAR,
-            cantidad_total DOUBLE,
-            valor_total DOUBLE,
+            cantidad_total DECIMAL(18,2),
+            valor_total DECIMAL(18,2),
             num_facturas INTEGER
         )
     """)
@@ -170,7 +170,7 @@ def build_duckdb(duckdb_path: str | Path) -> str:
     expected = [(MONTH_CURRENT, TOTAL_CURRENT), (MONTH_PREV, TOTAL_PREV)]
     for i, m in enumerate(monthly):
         exp_month, exp_total = expected[i]
-        diff = abs(m[1] - exp_total)
+        diff = abs(float(m[1]) - exp_total)
         if diff > 0.01:
             print(f"  ❌ DIFF {exp_month}: esperado=${exp_total:,.2f}, obtenido=${m[1]:,.2f} (diff=${diff:.2f})")
             ok = False
@@ -193,7 +193,7 @@ def build_duckdb(duckdb_path: str | Path) -> str:
     for t in top10:
         expected_sku = next((s for s in TOP_SKUS if s["cod_producto"] == t[0]), None)
         if expected_sku:
-            diff_val = abs(t[2] - expected_sku["valor_total"])
+            diff_val = abs(float(t[2]) - expected_sku["valor_total"])
             status = "✅" if diff_val < 0.01 else "❌"
             print(f"  {status} {t[0]}: ${t[2]:,.2f} (esperado ${expected_sku['valor_total']:,.2f}, diff=${diff_val:.2f})")
         else:
